# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UITRun API",
    "description": "This is an API document for UIT Run.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "trivonhan@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/user/update-first-login": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "[USER] update student ID and full name at first login",
        "description": "Multiple status values can be provided with comma separate strings",
        "operationId": "updateUserFirstLoginById",
        "requestBody": {
          "description": "Information",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "stdId": {
                    "type": "string"
                  },
                  "fullName": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "400": {
            "description": "This user has been updated.",
            "content": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "[ADMIN] get a user by id",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id to get",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "400": {
            "description": "Invalid queries value",
            "content": {}
          },
          "401": {
            "description": "Invalid Credentials.",
            "content": {}
          },
          "403": {
            "description": "You are not Admin. Permission Denied.",
            "content": {}
          },
          "404": {
            "description": "User doesn't exist.",
            "content": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "[ADMIN, EDITOR] Update an existing user by id.",
        "operationId": "updateUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "User info to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "400": {
            "description": "Invalid queries value",
            "content": {}
          },
          "401": {
            "description": "Invalid Credentials.",
            "content": {}
          },
          "403": {
            "description": "You are not Admin. Permission Denied.",
            "content": {}
          },
          "404": {
            "description": "User doesn't exist.",
            "content": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "[ADMIN] Deletes a user by id.",
        "operationId": "deleteUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {}
          },
          "400": {
            "description": "Invalid queries value",
            "content": {}
          },
          "401": {
            "description": "Invalid Credentials.",
            "content": {}
          },
          "403": {
            "description": "You are not admin. Permission Denied.",
            "content": {}
          },
          "404": {
            "description": "User doesn't exist.",
            "content": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "user": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string",
            "description": "fullname"
          },
          "stdId": {
            "type": "string",
            "description": "student id"
          },
          "userCode": {
            "type": "string",
            "description": "date"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}